% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.GPD.R
\name{dgpd}
\alias{dgpd}
\title{dgpd}
\usage{
dgpd(x, loc = 0, scale = 1, shape = 0, log = FALSE)
}
\arguments{
\item{x}{[vector] Vector of values}

\item{loc}{[vector] Location parameter}

\item{scale}{[vector] Scale parameter}

\item{shape}{[vector] Shape parameter}

\item{log}{[bool]   Return log of density if TRUE, default is FALSE}
}
\value{
[vector] Density of GPD at x
}
\description{
Density function of Generalized Pareto Distribution
}
\examples{
## Data
loc = 1
scale = 0.5
shape = -0.2
x = base::seq( -5 , 5 , length = 1000 )
y = dgpd( x , loc = loc , scale = scale , shape = shape )
}
