% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.GEV.R
\name{rgev}
\alias{rgev}
\title{rgev}
\usage{
rgev(n = 1, loc = 0, scale = 1, shape = 0)
}
\arguments{
\item{n}{[int]    Numbers of values generated}

\item{loc}{[vector] Location parameter}

\item{scale}{[vector] Scale parameter}

\item{shape}{[vector] Shape parameter}
}
\value{
[vector] Random value following a GEV(loc,scale,shape)
}
\description{
Random value generator of Generalized Extreme Value distribution
}
\examples{
## Data
loc = 1
scale = 0.5
shape = -0.2
gev = rgev( 100 , loc = loc , scale = scale , shape = shape )
}
