% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.GPD.R
\name{rgpd}
\alias{rgpd}
\title{rgpd}
\usage{
rgpd(n = 1, loc = 0, scale = 1, shape = 0)
}
\arguments{
\item{n}{[int]    Numbers of values generated}

\item{loc}{[vector] Location parameter}

\item{scale}{[vector] Scale parameter}

\item{shape}{[vector] Shape parameter}
}
\value{
[vector] Random value following a loc + GPD(scale,shape)
}
\description{
Random value generator of Generalized Pareto distribution
}
\examples{
## Data
loc = 1
scale = 0.5
shape = -0.2
gev = rgpd( 100 , loc = loc , scale = scale , shape = shape )
}
