% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.rorthogonal_group.R
\name{rorthogonal_group}
\alias{rorthogonal_group}
\title{rorthogonal_group}
\usage{
rorthogonal_group(d, n = 1)
}
\arguments{
\item{d}{[integer] Dimension of the matrix}

\item{n}{[integer] numbers of samples drawn}
}
\value{
[array or matrix], dim = d * d * n or d * d if n == 1
}
\description{
Generate sample from the orthogonal group O(d)
}
\examples{
M = ROOPSD::rorthogonal_group( 2 , 10 )

}
